requirements.txt
discord.py
Flask
psycopg2-binary
gunicorn
